Convert several modules to TypeScript
Fixes #143, Fixes #138, Fixes #137, Fixes #136

 * QueryResults has been converted to TypeScript
 * Filter has been converted to TypeScript
 * Store and tests/unit/Store have been converted.
 * Request and tests/unit/Request have been converted.

All test cases that were previously passing are still
passing, as well as the Request test cases.

QueryResults needed to be modified to return an ES6 style
promise, because of calls to super made by the promise
returned from dojo-core/Request. This conversion in turn
required many tests cases and some sources files to be
updated, as they were relying on the fact that Dojo 1 style
promises execute synchronously in some cases.