PR fixes
Removing unused code, pulling out common interfaces, using
existing interfaces(e.g. Hash) when applicable, and fixing
redundant logic in Filter.

Filter.js used to do call 'this.and.call(Filter.prototype, this, filter)',
which would and this and filter, however, prior to this
commit it's calling Filter.ts was calling
'this.and(this, filter)', which ands this with itself and
filter. Changing the call to this.and(filter) eliminates the
redundancy.